backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Futa",email = "fuukou160@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "google-api-python-client (>=2.170.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth (>=2.33.0,<3.0.0)",
    "google-cloud-secret-manager (>=2.23.0,<3.0.0)",
    "cryptography (>=42.0.5,<43.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "openai (>=1.83.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
